+++ packages/api/src/hooks/index.ts
@@ -12,6 +12,7 @@ export { default as useGetAccountStatus } from './useGetAccountStatus';
 export { default as useLandingCompany } from './useLandingCompany';
 export { default as useMT5AccountsList } from './useMT5AccountsList';
 export { default as useSettings } from './useSettings';
+export { default as useSortedMT5Accounts } from './useSortedMT5Accounts';
 export { default as useTradingAccountsList } from './useTradingAccountsList';
 export { default as useTradingPlatformAccounts } from './useTradingPlatformAccounts';
 export { default as useTradingPlatformAvailableAccounts } from './useTradingPlatformAvailableAccounts';
