+++ packages/wallets/src/components/AddedMT5AccountsList/AddedMT5AccountsList.scss
@@ -1,0 +1,90 @@
+.wallets-added-mt5 {
+    &__transfer_button {
+        display: flex;
+        height: 32px;
+        padding: 6px 16px;
+        justify-content: center;
+        align-items: center;
+        border-radius: 4px;
+        border: 1px solid var(--system-light-3-less-prominent-text, #999);
+        cursor: pointer;
+        background: #fff;
+    }
+
+    &__transfer_text {
+        color: var(--system-light-1-prominent-text, #333);
+        text-align: center;
+
+        /* desktop/paragraph/P2 - bold */
+        font-family: 'IBM Plex Sans';
+        font-size: 14px;
+        font-style: normal;
+        font-weight: 700;
+        line-height: 20px; /* 142.857% */
+    }
+
+    &__open_text {
+        color: var(--system-dark-1-prominent-text, #fff);
+        text-align: center;
+
+        /* desktop/paragraph/P2 - bold */
+        font-family: 'IBM Plex Sans';
+        font-size: 14px;
+        font-style: normal;
+        font-weight: 700;
+        line-height: 20px; /* 142.857% */
+    }
+
+    &__actions {
+        display: flex;
+        flex-direction: column;
+        gap: 4px;
+    }
+
+    &__details {
+        flex-grow: 1;
+
+        &-loginid {
+            color: var(--system-light-3-less-prominent-text, #999);
+
+            /* desktop/small/S - bold */
+            font-family: 'IBM Plex Sans';
+            font-size: 12px;
+            font-style: normal;
+            font-weight: 700;
+            line-height: 18px; /* 150% */
+        }
+
+        &-title {
+            color: var(--system-light-2-general-text, #333);
+
+            /* desktop/paragraph/P2 - bold */
+            font-family: 'IBM Plex Sans';
+            font-size: 14px;
+            font-style: normal;
+            font-weight: 700;
+            line-height: 20px; /* 142.857% */
+        }
+
+        &-balance {
+            color: var(--system-light-1-prominent-text, #333);
+
+            /* desktop/paragraph/P2 - bold */
+            font-family: 'IBM Plex Sans';
+            font-size: 14px;
+            font-style: normal;
+            font-weight: 700;
+            line-height: 20px; /* 142.857% */
+        }
+
+        &-description {
+            font-size: 1.2rem;
+            line-height: 14px;
+
+            @include mobile {
+                font-size: 1.4rem;
+                line-height: 20px;
+            }
+        }
+    }
+}
