+++ packages/wallets/src/components/AddedMT5AccountsList/index.ts
@@ -1,0 +1,1 @@
+export { default as AddedMT5AccountsList } from './AddedMT5AccountsList';
