+++ packages/wallets/src/components/AvailableMT5AccountsList/AvailableMT5AccountsList.scss
@@ -1,0 +1,32 @@
+.wallets-available-mt5 {
+    &__text {
+        color: var(--brand-coral, #ff444f);
+        text-align: center;
+
+        /* desktop/paragraph/P2 - bold */
+        font-family: 'IBM Plex Sans';
+        font-size: 14px;
+        font-style: normal;
+        font-weight: 700;
+        line-height: 20px; /* 142.857% */
+    }
+
+    &__details {
+        flex-grow: 1;
+        &-title {
+            font-weight: bold;
+            font-size: 1.4rem;
+            line-height: 20px;
+        }
+
+        &-description {
+            font-size: 1.2rem;
+            line-height: 14px;
+
+            @include mobile {
+                font-size: 1.4rem;
+                line-height: 20px;
+            }
+        }
+    }
+}
