+++ packages/wallets/src/components/AvailableMT5AccountsList/index.ts
@@ -1,0 +1,1 @@
+export { default as AvailableMT5AccountsList } from './AvailableMT5AccountsList';
