+++ packages/wallets/src/components/index.ts
@@ -1,7 +1,9 @@
 export * from './AccountsList';
 export * from './AddedDxtradeAccountsList';
+export * from './AddedMT5AccountsList';
 export * from './AvailableDerivezAccountsList';
 export * from './AvailableDxtradeAccountsList';
+export * from './AvailableMT5AccountsList';
 export * from './CFDList';
 export * from './CTraderList';
 export * from './DesktopWalletsList';
